ENH: Support passing in a template model hierarchy file to the Model Maker
Assumption is that the provided .mrml file has been created to match the
color information associated with the input label  map volume. Models
are created and added to the hierarchy as associated nodes of the
model hierarchy nodes that are named for the color of the label that
the model was created from.
Cleans up the output scene to remove unused leaf hierarchy nodes.

TODO: create and commit .mrml files for the standard anatomy based color nodes