ENH: on click in the name column, jump slices to that point location
Use the logic methods to jump the slices after extracting
the location from the item clicked on.
For now, clicking in the name column triggers the jump slices.
Set selection mode to row and allow multi select for future
mass operations.

TODO: make a new column to click in for jump or implement righ click
on the row.