BUG: add a display node for a fiducial node created via the combo box
React to the user created signal from the combo box and use the logic
to create and add a display node. Previously the newly created fiducials
were appearing incorrectly in the displayable managers because they didn't
have display nodes.
Update the test to use the renamed logic method, want to create
a fiducial node rather than a generic markups node.