BUG: clear out the table when the last markups node is removed
When a user deletes nodes via the node selector, the table was
updating correctly to a remaining node until the last node
was removed. This change checks on a node removed event
if there are not markups nodes left in the scene and if
that's the case, it clears out the GUI.
Added clearGUI as a utility method.