ENH: add the ability to add custom actions and trigger them
Added a list of actions that are added after the default none,
add, remove, rename actions that the combo box supplies. When
the new action is selected, it is triggered so a module can
add a connection to listen for the triggered signal on the action
and activate a local slot.
Added a test to exercise trying to add duplicate or conflicting
actions.