ENH: add persistent saving of default display node properties
Use the ctk settings on the application to save default
markups display node properties across executions of Slicer.
TODO: debug why the settings panel interacts poorly with the saved
settings in the logic class.
Added a button to save defaults in the GUI.
Added testing.
Added an event so that a markup added via mouse click will have
it's display node set up from the defaults.