ENH: respond to end batch process events to update widget
When a markups file is loaded, the node added event is ignored
because the scene is in batch process mode. Connect and respond
to the end batch process event so that when a markups file
is loaded and it's the first markup in the scene, it will be
set active and shown in the GUI. If the scene contains an
active markup node already, the newly loaded one will not
be set active, it has to be selected from the drop down
menu.

STYLE: removed trailing spaces.