BUG: prevent MRML updates to the table from updating MRML
Blocking signals is a brute force solution, but it's preventing
the seed widget from getting further and further from the mouse
as a fiducial is dragged. In updateRow updating the cells with
the fiducial location was triggering a cell changed signal
even with checking that the number value as a text string
had changed.