BUG: fix jumping of all fiducials to the slice where one fid was
The widget call backs for 2d and 3d fiducial displayable managers had
the same class name and was causing a conflict and resetting of
mrml coordinates.
There was also a logic bug that was calling UpdateNthMarkupPosition
when the seed widget position changed, but that was updating
the seed position. Renamed that method to UpdateNthSeedPositionFromMRML
to be very clear about what it does, added UpdateNthMarkupPositionFromWidget
and call it from the call back.
Updated the check on if positions changed when updating all the points
to be more logically clear. TODO: check if need the disable modified
and explicit invokation of point location modified/markups modified.