ENH: add ability to add a fiducial when press the p key
When in markups fiducial place mode and the markups gui is up,
the user can use the p key to place a new fiducial. Synthesize
a mouse click and pass it to the widget under the mouse cursor.

TODO: there's a crash when setting the shortcut parent to null
and try placing a fiducial via mouse click when not in the markups
module. Need to get the shortcut really disabled, as of now, it's
just disconnected from the slot that will pass the event
to the widgets.