ENH: add ability to move a markup to another list
Added a right click context menu with a sub menu to select where
in the new list to move the markup, then pop up a dialog
to allow the user to select the target list. Only works on
one row at a time for now. Use the object name on the
list selection dialog to pass along the target position.
Added logic and node methods to move and insert markups
in the markups node. Added a new logic test for the move
functionality.