ENH: make click to jump optional
Added a checkbox to enable clicking in the name column
triggering a jump to slices call in the slice viewers.
Added two radio buttons to choose if the jump should be
offset or centered.
From feedback at the MIT Summer '13 project week,
defaults set to no click to jump, centered.