ENH: add a unique point id to the markups
Keep track of the number of markups that were added to the list and use that
max number for labels as well as the id if the scene is not available to generate a
unique ID. Reset the maximum number when all markups are removed, but don't reduce it
when individual markups are deleted from the list.

Simplified the default node names to start with F so that point names are shorter.

Updated the storage node to print out the IDs and read them in, removed some old code, adding testing.