BUG: add support for LPS and IJK points to CLIs
Add an optional coordinate system flag to the WriteCLI method on
displayable nodes. Check the coordinate system parameter in the
CLI module logic and set the flag when calling WriteCLI to write
out points. Use the flag in the Annotations control point node
to write out LPS if required, IJK is not supported yet (writes
out RAS as the default).

Issue #2979