ENH: simplified the storage node, added LPS support
Made a fiducials specific storage node and simplified the file format.
Added support for outputting points in LPS or RAS in the storage file
or on the CLI command line.
Added tests to ensure that can read Slicer3 fcsv files and removed
the registration of that IO factory in the annotations module.
Slicer3 fids will be read in as markup fiducials now.