BUG: deal with commas and quotes in markup label and description
Added utility methods on the nodes to put quotes around the label
and description strings before writing to file, and double quotes
for internal quotes in the string. Convert back to non storage
format for display in the UI. There should be no disallowed
characters in the label or description, so no entry
validator needed.
Added parsing using the quoted strings in the storage node.
Added testing.