BUG: debugging locking of widgets when leave place mode
The state of the interaction node needed to be checked when
locking seeds, so pass it into the helper.
Added right click in 3d viewer to leave place mode (stays
persistent though), for both Markups and Annotations (adding
that exposed the bug in locking, as the markups weren't
being unlocked properly).
Right click in 2d is possibly being swallowed by the
ctkQImageView zoom.

Issue #3211, #2717