BUG: get the unit tests to show up in the SelfTests module
Build with the slicerMacroBuildScriptedModule so that the .py file
is copied into the scripted modules directory.
Currently, it seems like the slicer_add_python_unittest macro is also
needed so that the test can be run via ctest.
Now the add many fiducials and the neuro tutorial test show up in
the self tests module.