ENH: add markup label format string
Instead of relying on just the list name, let the user specify
a sprintf style format string to set the markup labels.
Added testing.
Added buttons to reset the format to default, and to rename
all markups in the list.
Also implemented the table disable when the list is locked.

Todo: fix the generic widget test again.