BUG: bug fixing and tweaks from user feedback
Fix pan bug, comparison for valid viewpoint coordinates wasn't properly
checking the first coordinate (fixed it in Annotations as well).
Replace Annotations with Markups in default favorite modules list.
Add link to wiki help page
Updated default glyph and text size as per Ron's suggestion
Fix the bug where a double click on the visibility, lock or selected could
pop an entry box (make then not editable).
Adjust the cell selection in the table so that it works with mouse clicks
as well as keyboard arrows
Update the Endoscopy module to be able to use Markup fiducials (added a
backward compatibility call of GetNumberOfFiducials to the Markups fiducial node).
Added a compare view self test.
Take out changing file paths for testing files to native format,
was causing a compile error on Windows.
Add in an empty category so that the Markups module will appear
at the top of the modules drop down as well as in Informatics.
Changed default color to match Slicer3 where fiducials were created with selected
on by default.
Removed using markups node name in name format string, user can use %N manually
Added toggle methods for all markups in list for visibility, locked, selected
and consolidated the calls into menu push buttons in the GUI (added a select/unselect icon)
Fixed a bug in deleting projections, have to check all points in lists against
all the projection widgets since the widget delete method was being called
after a point was removed from the list and the projection widget was no
longer mapped to a valid point id.
Disable projection color picker button if using fiducial color.
Added a uniform markup scale that will set both text and glyph to the same scale.
Adjusted table column widths (smaller) and RAS coordinate precision (to 3 from 6)
Factored out the GUI update to watch for and respond to display node changes.
Added a jump slices action to the right click menu in the table.
Change fiducial place icon to have arrow point at top, the previous icon had the
fiducial glyph at the top of the icon and was leading to a feeling that the
fiducial was off by a few pixels when placed with the mouse. Using a copy
of the Annotations point with arrow icon.
Update the XML description to match the new GUI layout, adding
new entries for new widgets and updating some descriptions
from ui file tool tips.