BUG: bug fixing from user feedback - pan, click, help, endoscopy
Fix pan bug, comparison for valid viewpoint coordinates wasn't properly
checking the first coordinate (fixed it in Annotations as well).
Replace Annotations with Markups in default favorite modules list.
Add link to wiki help page
Updated default glyph and text size as per Ron's suggestion
Fix the bug where a double click on the visibility, lock or selected could
pop an entry box (make then not editable).
Adjust the cell selection in the table so that it works with mouse clicks
as well as keyboard arrows
Update the Endoscopy module to be able to use Markup fiducials (added a
backward compatibility call of GetNumberOfFiducials to the Markups fiducial node).
Added a compare view self test.