ENH: add copy to markups list right click option
Can now right click on a markup in the GUI and copy it
to another list, placing it at the end of the destination
list.
Updated the right click Move option to only offer at the
end of the list as that simplifies the UI and users can
move the markups with the move up/down buttons.
Updated test for logic copy.
Updated documentation for copy.

Issue #3397