ENH: add active manipulation node attributes, selected view visibility
Porting from trunk.

Tell which slice viewer is manipulating which fiducial in the
markup node.

Markups.MovingInSliceView holds the LayoutName of the sliceNode associated
with the viewer where the widget is being manipulated.

Markups.MovingMarkupIndex holds the seed number being manipulated.

Both are cleared when the seed is not being manipulated by a widget.

Give annotations option to be visible only in selected views

This uses the same convention used in dislay nodes for other
displayable managers (e.g. VolumeRendering), except that it uses the
sliceNode since it is the 2D analog to the 3D view Node (at some point
perhaps there should be a generic superclass).  Here, since we are only
using the ID string it doesn't matter what type the node it.

This feature allows custom UIs, for example where some annotations only
show up in one view while others show up everywher.  In particular, this
is helpful for annotations representing volume-specific anatomical
landmark locations.