ENH: add coordinates to right click menu
When select rows in the Markups table, right click will now add the
coordinate points to the top of the context menu. Works for
multi select, and will work for multiple points per markup.
Adds the label of the markup if there are multiple rows selected.
Adds a separator after the coordinates and before the command
actions.