ENH: adjust the scale slider widgets to volume slice spacing
On entry to the Markups module, check for a background volume
in the red slice viewer, get the slice spacing and adjust
the scale slider maximum values by a scale factor times
the maximum slice spacing.
Also ensure that if the module is entered with scale values
on the display node that are larger than the slider maximums
that the sliders are adjusted.
Also works for the default settings being larger than slider
maximums.

Issue #3547