ENH: Expand scalar range options
Add the ScalarRangeFlag on the display node, test it, use it in the Model
display node.
Provide a GUI interface for it in the Models Scalars GUI.
Use it in the model displayable manager.
Separate out the auto scalar range setting from updating the poly data
in the model display node.

This expands the options from just using the scalar range of the data
to allowing the user to select using the scalar range defined on the
display node, the range on the color node, or the range of the data
type (use this last with caution, TBD: readjusting the minimum and
maximum values allowed on the range slider). Keeps the default
behaviour of AutoScalarRange and Data. The interaction of Auto/Manual
and the range flags might need some tweaking.
Also keeps the deep copy of the color look up table in the model displayable
manager - monitor for performance issues.
Expands the model displayable manager to work with procedural color nodes,
this required copying the color transfer function to a look up table to
support setting the range. Using a table with 256 entries for now.

Issue #3573