ENH: respond to MRML events, expand functionality.
Watch for and respond to the scene closing, restoring, batch processing.
Clear out the temp files on scene close and clear the memory cache, now
don't have to write them out all the time. Resize the images before
writing to disk so that the web view doesn't have to handle resizing.
Add ability to create new and delete scene views.
Link the description to editing the scene view, but make it a hidden
link (no underline, black text).
Remove explicit update of tree view in preparation to removing
the UI elements.

Issue #1299