ENH: scroll bar as needed
Running on a mac, you can't see the scroll bar
on the data store page nor on the scene views
module widget, but it is visible on a linux build.
You can scroll using the trackpad and it is consistent,
so change to only show the scroll bar as needed.

Issue #1299