ENH: use the scalars to colors annotation with scalar bar
Initial test implementation generating the same number of annotations
as labels.

Issues:
Changing the number of labels on the ctk widget doesn't trigger a modified event, so
don't have the signal to update the annotations manually. Workaround: toggle
visibility.
Should be deep copying the color look up table rather than changing the Slicer singleton.
Should we annotate all the color table node lookup tables? Then would need to subclass
the scalar bar actor to get the color names in the labels position and not always show all
of the color names. Left the stub for color names in the labels field, not implemented yet.

Issue #2919