ENH: update GUI to show second ruler point
Add support for columns in the GUI to show the second
ruler point and to switch back and forth with one point
columns when have a fiducial list.
Fixed a bug with showing the transformed coordinates.
Generalised the GUI some more for copying and moving
markups between lists (make sure only move between the
same types of lists for simplicity).

Issue #3367