ENH: add formatting for ruler distance
Add the Slicer 3 options to select some default formats for how
the distance is shown on the ruler. Uses the Annotations module ruler default.

Adds the distance measurement format to the ruler display node, tests it.
Adds some default format settings in a combo box and a free form line
edit widget to show/set the format in the GUI.

Issue #3367