BUG: update cursor on persistent place type change
When in persistent place mode, a user can change to a different
type of node to place without dropping out of place mode. This
change fixes the bug that the cursor won't get updated because
the interaction mode change is a no-op and a modified event
doesn't get triggered to update the tool bar. This is a bit of
a hack, manually updating the tool bar to place mode to trigger
a cursor change to the correct icon.