BUG: fix crash when showing transformed markups
The module widget was getting the world coordinates by passing in a 3
element array, not a 4 element array. On some machines this was
triggering a crash. This fix avoids that by properly passing
a four element point array and then reassigning the values to
the three element array for display.

Issue #3756