BUG: consolidate distance from fid to slice plane
Fiducials are projected onto the slice plane for display, showing up
if they're within some small delta from the plane. Trying to stop the fiducial
from jumping onto the plane on mouse over and drag.
Consolidate the distance calculcation into one method and use it in the
displayable managers.

Issue #3531