BUG: fix scaling between 2d and 3d seed widgets
WIP:
Added calculating a better camera viewing position for the 2d slice view
cameras from the data, or if no data, use the 3d camera's position. This
is a change from the default (not used by anything other than the VTK widgets)
of a viewing distance of 1, which resulted in needing a scale factor.
Removed the scaling factor in the 2d displayable manager.
Updated GetLightboxIndex to return 0 on failure so that it will lead to
getting the first light box manager's renderer.
Change GetRenderer calls to get the 0th one from the light box renderer
manager.
Observe the slice composite node for background volume changes to trigger
a camera distance update.
TBD: restore seed interaction, fix disappearing fids when data is loaded,
fix memory leaks.

Issue #3547