BUG: update fiducial layout test to check RAS
The former version of this test was set to always pass
since there was still a > 1 pixel difference in starting
and ending fiducial display coordinates. Brainstorming
on the slicer dev hangout brought up the possibility
that since the fiducial RAS position was still correct,
the layout switching might have included some FOV adjustments
to fit the volume to the slice widget. This change will
go on to check the fiducial RAS position against the
volume RAS position if the display positions don't match.

Update the layout switching test to:
- fail on the dashboard if it fails
- show all output for debugging
- check both display coordinates and RAS coordinates
- only fail if both are out of tolerance
- remove the redundant delayDisplay methods and use the slicer.util
version with a shorter time out to speed up the test run time

Issue #1914