ENH: check for data that will get deleted on scene view restore
Added a flag to the scene view node restore call that will cancel if
any nodes come up as being in the main Slicer scene but not in the
scene view scene about to be restored.
Added a method to add nodes to the scene view scene that are in the
main Slicer scene but not in the scene view.
Updated the GUI for restoring scenes to pop up a message box
to give the user options to discard the data, save it to the scene
view, or cancel the restore.
Updated the test to include a failure case when removing nodes
is not allowed.

Issue #3956