BUG: don't overwrite model storage node ids when reading scalar overlays
In the models logic AddScalar, don't add the storage node ID to the model node
manually, let the storage node read do it automatically.
In vtkMRMLStorageNode::ReadData change the setting of the storage node id call to
an add in order to not over write the model storage node when adding scalar overlays.

Add a check in vtkMRMLNode::AddAndObserveNodeReferenceID to not add the same id in the same role.
TBD: should an update be called instead of just returning null?

Issue #3684